updated demo and fixed a int->float conversion in embed

+++ js/plain/quick_graph_embed.js
@@ -80,7 +80,7 @@ function returnVariableMinStepMax(embeddedGraph,variableNumber,MSM,refNum,attrib
     {    
         var triplets = minstepmaxString.split(";");
         var singlets = triplets[variableNumber].split(",");
-        return parseInt(singlets[MSM]);
+        return parseFloat(singlets[MSM]);
     }
     else
     {
